PollInput yazildi (park, 4 bayt); uc yeni kol olculdu
0x080656F4 (228 B): tus okuma ve giris gecmisi halka tamponu. Omuz
tuslari basili degilken dort yon birden basiliysa donanim kapanisi.
Baglanti durumu 2 iken her kare 32 elemanlik halka tampona yaziliyor.

185 -> 4 bayt; uc kol bulundu ve dosyada kayitli:
1. Ayrik maskeli iki tus testi AYRI `if` olmali; tek `&&` ifadesinde GCC
   0x300 ve 0x00F'i tek maskeye (0x30F) katliyor.
2. `index` s32 olmali; u16 bildirimi her atamada 16 bit kirpma ekliyor.
3. Bayt daraltmasi `& 0xFF` olmali, `(u8)` cast'i degil: cast 24 bit
   kaydirma cifti uretip fazladan callee-saved yazmac tuketiyor
   (dump_alloc pseudo 119/120/121).

Permuter 1296 yinelemede 875 -> 425, sifir yok. En iyi aday `index`i
`unsigned long long` yapiyor; u16 sayac icin gercek kaynak olamaz,
reddedildi.

include/game_state.h'ye +0x08 ve +0x0A giris alanlari, gba_io.h'ye
REG_KEYINPUT eklendi; mevcut eslesmeler korundu.

+++ src/world/link_session_reset.c
@@ -28,7 +28,7 @@ typedef struct LinkCounters {
 
 extern CommBlock   gLinkBlock;
 extern LinkCounters gRam02036330;
-extern u8  gRam02000420[];
+extern u16 gRam02000420[];
 extern u8  gRam02000230[];
 extern u8  gRam020003C0[];
 extern u8  gRam02000E80[];
